+++ src/all/koharu/src/eu/kanade/tachiyomi/extension/all/koharu/Koharu.kt
@@ -1,15 +1,7 @@
 package eu.kanade.tachiyomi.extension.all.koharu
 
-import android.annotation.SuppressLint
 import android.app.Application
 import android.content.SharedPreferences
-import android.os.Handler
-import android.os.Looper
-import android.util.Log
-import android.webkit.WebResourceRequest
-import android.webkit.WebResourceResponse
-import android.webkit.WebView
-import android.webkit.WebViewClient
 import androidx.preference.ListPreference
 import androidx.preference.PreferenceScreen
 import androidx.preference.SwitchPreferenceCompat
@@ -30,7 +22,6 @@ import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
 import okhttp3.HttpUrl.Companion.toHttpUrl
 import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.Response
@@ -38,11 +29,8 @@ import rx.Observable
 import uy.kohesive.injekt.Injekt
 import uy.kohesive.injekt.api.get
 import uy.kohesive.injekt.injectLazy
-import java.io.IOException
 import java.text.SimpleDateFormat
 import java.util.Locale
-import java.util.concurrent.CountDownLatch
-import java.util.concurrent.TimeUnit
 
 class Koharu(
     override val lang: String = "all",
@@ -62,15 +50,6 @@ class Koharu(
     private val authUrl = "${baseUrl.replace("://", "://auth.")}/clearance"
 
     override val supportsLatest = true
-
-    override val client: OkHttpClient = network.cloudflareClient.newBuilder()
-        .rateLimit(5)
-        .build()
-
-    private val interceptedClient: OkHttpClient = network.cloudflareClient.newBuilder()
-        .addInterceptor(::cloudflareTurnstileInterceptor)
-        .rateLimit(5)
-        .build()
 
     private val json: Json by injectLazy()
 
@@ -113,6 +92,16 @@ class Koharu(
             .set("Origin", domainUrl)
             .build()
     }
+
+    override val client: OkHttpClient = network.cloudflareClient.newBuilder()
+        .rateLimit(5)
+        .build()
+
+    private val interceptedClient: OkHttpClient
+        get() = network.cloudflareClient.newBuilder()
+            .addInterceptor(TurnstileInterceptor(client, domainUrl, authUrl, lazyHeaders["User-Agent"]))
+            .rateLimit(5)
+            .build()
 
     private fun getManga(book: Entry) = SManga.create().apply {
         setUrlWithoutDomain("${book.id}/${book.key}")
@@ -410,204 +399,6 @@ class Koharu(
         return data.similar.map(::getManga)
     }
 
-
-    /* Cloudflare Turnstile interceptor */
-
-    private fun authHeaders(authorization: String) =
-        headersBuilder()
-            .set("Referer", "$domainUrl/")
-            .set("Origin", domainUrl)
-            .set("Authorization", authorization)
-            .build()
-
-    private val authorizedRequestRegex by lazy { Regex("""(.+\?crt=)(.*)""") }
-
-    fun cloudflareTurnstileInterceptor(chain: Interceptor.Chain): Response {
-        if (token == null) {
-            resolveInWebview()
-        }
-        val request = chain.request()
-
-        val url = request.url.toString()
-        Log.e("Koharu", "Requesting URL: $url")
-        val matchResult = authorizedRequestRegex.find(url) ?: return chain.proceed(request)
-        if (matchResult.groupValues.size == 3) {
-            val requestingUrl = matchResult.groupValues[1]
-            val crt = matchResult.groupValues[2]
-            var newResponse: Response
-
-            if (crt.isNotBlank() && crt != "null") {
-                // Token already set in URL, just make the request
-                newResponse = chain.proceed(request)
-                Log.e("Koharu", "Response code: ${newResponse.code}")
-                if (newResponse.code !in listOf(400, 403)) return newResponse
-            } else {
-                // Token doesn't include, add token then make request
-                if (token.isNullOrBlank()) resolveInWebview()
-                val newRequest = if (request.method == "POST") {
-                    POST("${requestingUrl}$token", lazyHeaders)
-                } else {
-                    GET("${requestingUrl}$token", lazyHeaders)
-                }
-                Log.e("Koharu", "New request: ${newRequest.url}")
-                newResponse = chain.proceed(newRequest)
-                Log.e("Koharu", "Response code: ${newResponse.code}")
-                if (newResponse.code !in listOf(400, 403)) return newResponse
-            }
-            newResponse.close()
-
-            // Request failed, refresh token then try again
-            clearToken()
-            resolveInWebview()
-            val newRequest = if (request.method == "POST") {
-                POST("${requestingUrl}$token", lazyHeaders)
-            } else {
-                GET("${requestingUrl}$token", lazyHeaders)
-            }
-            Log.e("Koharu", "New re-request: ${newRequest.url}")
-            newResponse = chain.proceed(newRequest)
-            Log.e("Koharu", "Response code: ${newResponse.code}")
-            if (newResponse.code !in listOf(400, 403)) return newResponse
-            throw IOException("Solve Captcha in WebView (${newResponse.code})")
-        }
-        return chain.proceed(request)
-    }
-
-    private val context: Application by injectLazy()
-    private val handler by lazy { Handler(Looper.getMainLooper()) }
-
-    @SuppressLint("SetJavaScriptEnabled")
-    fun resolveInWebview(): Pair<String?, String?> {
-        Log.e("TurnstileInterceptor", "resolveInWebview")
-        val latch = CountDownLatch(1)
-        var webView: WebView? = null
-        var tokenRequested = false
-
-        handler.post {
-            val webview = WebView(context)
-            webView = webview
-            with(webview.settings) {
-                javaScriptEnabled = true
-                domStorageEnabled = true
-                databaseEnabled = true
-                useWideViewPort = false
-                loadWithOverviewMode = false
-            }
-
-            webview.webViewClient = object : WebViewClient() {
-                override fun shouldInterceptRequest(view: WebView?, request: WebResourceRequest?): WebResourceResponse? {
-                    val authHeader = request?.requestHeaders?.get("Authorization")
-                    if (request?.url.toString().contains(authUrl) && authHeader != null) {
-                        authorization = authHeader
-                        if (request.method == "POST") {
-                            Log.e("TurnstileInterceptor", "Authorization: $authorization")
-                            tokenRequested = true
-
-                            try {
-                                val noRedirectClient = client.newBuilder().followRedirects(false).build()
-                                val authHeaders = authHeaders(authHeader)
-                                val response = noRedirectClient.newCall(POST(authUrl, authHeaders)).execute()
-                                response.use {
-                                    if (response.isSuccessful) {
-                                        with(response) {
-                                            token = body.string()
-                                                .removeSurrounding("\"")
-                                            Log.e("TurnstileInterceptor", "Requested token: $token")
-                                        }
-                                        latch.countDown()
-                                    } else {
-                                        println("Request failed with code: ${response.code}")
-                                    }
-                                }
-                            } catch (e: IOException) {
-                                println("Request failed: ${e.message}")
-                                latch.countDown()
-                                return super.shouldInterceptRequest(view, request)
-                            }
-                        }
-                        if (request.method == "GET") {
-                            // TODO: What to do if it return failed? => should not countdown, let the POST method request again
-                            // Can try mitigate a wrong clearance
-                            // Token might be rechecked here but just let it fails then we will reset and request a new one
-                            // Normally this might not occur because old token should already be returned via onPageFinished
-                            Log.e("TurnstileInterceptor", "Authorization: $authorization")
-                            token = authorization?.substringAfterLast(" ")
-                            tokenRequested = true
-                            latch.countDown()
-                        }
-                    }
-                    return super.shouldInterceptRequest(view, request)
-                }
-
-                override fun onPageFinished(view: WebView?, url: String?) {
-                    if (view == null) return
-                    // Read the saved token in localStorage
-                    // Fixme: this might overwrite the one newly requested
-                    val script = "javascript:localStorage['clearance']"
-                    view.evaluateJavascript(script) {
-                        if (!it.isNullOrBlank() && it != "null") {
-                            token = it
-                                .removeSurrounding("\"")
-                            Log.e("TurnstileInterceptor", "Clearance: $token")
-                            latch.countDown()
-                        }
-                        Log.e("TurnstileInterceptor", "Page finished")
-                    }
-                }
-            }
-
-            webview.loadUrl("$domainUrl/")
-        }
-
-        latch.await(20, TimeUnit.SECONDS)
-
-        handler.post {
-            if (token.isNullOrBlank()) {
-                val script = "javascript:localStorage['clearance']"
-                webView?.evaluateJavascript(script) {
-                    if (!it.isNullOrBlank() && it != "null") {
-                        token = it
-                            .removeSurrounding("\"")
-                    }
-                    Log.e("TurnstileInterceptor", "Clearance: $it / $token - Authorization: $authorization")
-                }
-            }
-
-            webView?.stopLoading()
-            webView?.destroy()
-            webView = null
-        }
-
-        return token to authorization
-    }
-
-    @SuppressLint("SetJavaScriptEnabled")
-    private fun clearToken() {
-        val latch = CountDownLatch(1)
-        handler.post {
-            val webView = WebView(context)
-            with(webView.settings) {
-                javaScriptEnabled = true
-                domStorageEnabled = true
-                databaseEnabled = true
-            }
-            webView.webViewClient = object : WebViewClient() {
-                override fun onPageFinished(view: WebView?, url: String?) {
-                    if (view == null) return
-                    val script = "javascript:localStorage.clear()"
-                    view.evaluateJavascript(script) {
-                        token = null
-                        view.stopLoading()
-                        view.destroy()
-                        latch.countDown()
-                    }
-                }
-            }
-            webView.loadUrl(domainUrl)
-        }
-        latch.await(20, TimeUnit.SECONDS)
-    }
-
     // Settings
 
     override fun setupPreferenceScreen(screen: PreferenceScreen) {
@@ -638,7 +429,7 @@ class Koharu(
         private const val PREF_IMAGERES = "pref_image_quality"
         private const val PREF_REM_ADD = "pref_remove_additional"
 
-        private var token: String? = null
-        private var authorization: String? = null
+        internal var token: String? = null
+        internal var authorization: String? = null
     }
 }
